+++ ampd/src/solana/mod.rs
@@ -1,4 +1,4 @@
-pub mod ws_verifier;
+pub mod verifier_set_verifier;
 pub mod msg_verifier;
 pub mod rpc;
